backend/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fsspec==2024.5.0
gradio_client==0.17.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
iniconfig==2.0.0
jedi==0.19.1
Jinja2==3.1.4
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ollama==0.2.0
opencv-python==4.9.0.80
orjson==3.10.3
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.45
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==11.0.3
